descriptions/requirements.txt
pandas>=2.0
yfinance>=0.2
requests>=2.28
sqlalchemy>=2.0
psycopg2-binary>=2.9
pyyaml>=6.0
